Task 7 Unit tests
