+++ src/main.ts
@@ -1,44 +1,10 @@
-import express from 'express';
-import cors from 'cors';
-import dotenv from 'dotenv';
-import cookieParser from 'cookie-parser';
-
+import app from './app.js';
 import db from './models/index.js';
-import { userRouter, groupRouter } from './routes/index.js';
-import { errorWinstonLogger, infoWinstonLogger, logger } from './utils/loggers.js';
-import errorHandler from './middleware/errorHandler.js';
-import checkAuthorization from './middleware/checkAuthorization.js';
+import { logger } from './utils/loggers.js';
 
 const { sequelize } = db;
-dotenv.config();
 
-const app = express();
 const port = process.env.PORT || 3030;
-
-app.use(express.json());
-app.use(cookieParser());
-app.use(
-  cors({
-    origin: '*',
-  }),
-);
-
-app.use(checkAuthorization);
-
-// express-winston logger makes sense BEFORE the router
-app.use(infoWinstonLogger);
-
-app.use(userRouter);
-app.use(groupRouter);
-
-app.use(errorHandler);
-
-// express-winston errorLogger makes sense AFTER the router
-app.use(errorWinstonLogger);
-
-app.get('/', (req, res) => {
-  res.send('Main Page');
-});
 
 sequelize
   .sync({ force: false })
