+++ src/models/index.js
@@ -13,7 +13,7 @@ import group from './group.js';
 import userGroup from './userGroup.js';
 
 const __filename = fileURLToPath(import.meta.url);
-const env = process.env.NODE_ENV || 'development';
+const env = process.env.NODE_ENV;
 const __dirname = path.dirname(__filename);
 const config = fs.readFileSync(path.resolve(__dirname, '../data-access/config.json'));
 const currentEnv = JSON.parse(config)[env];
