+++ src/routes/groupRoute.ts
@@ -1,12 +1,12 @@
 import { Router } from 'express';
 import GroupController from '../controllers/groupController.js';
 
-const userRouter = Router();
+const groupRouter = Router();
 
-userRouter.get('/api/group', GroupController.getAllGroups);
-userRouter.get('/api/group/:id', GroupController.getGroup);
-userRouter.post('/api/group', GroupController.createGroup);
-userRouter.put('/api/group', GroupController.updateGroup);
-userRouter.delete('/api/group/:id', GroupController.deleteGroup);
+groupRouter.get('/api/group', GroupController.getAllGroups);
+groupRouter.get('/api/group/:id', GroupController.getGroup);
+groupRouter.post('/api/group', GroupController.createGroup);
+groupRouter.put('/api/group', GroupController.updateGroup);
+groupRouter.delete('/api/group/:id', GroupController.deleteGroup);
 
-export default userRouter;
+export default groupRouter;
